### 2024-03-06 #6449 @editorialbot
[PRE REVIEW]: ADIOS4DOLFINx: A framework for checkpointing in FEniCS
**Submitting author:** @jorgensd (<a href="http://orcid.org/0000-0001-6489-8858">Jørgen Schartum Dokken</a>)
**Repository:** https://github.com/jorgensd/adios4dolfinx
**Branch with paper.md** (empty if default branch):
**Version:** v0.8.0
