### 2024-03-06 #6449 @danielskatz
👋 @jorgensd - thanks for your submission. I'll be editing it.

If you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). You can search [the list of people](https://reviewers.joss.theoj.org/lookup) that have already agreed to review and may be suitable for this submission. You can also suggest people not on the JOSS list.
