### 2024-03-07 #6449 @danielskatz
Yes, folks that you have a conflict of interest with can't review. Note that our COI rule goes back 4 years on previous collaborations, so if something is older than that, it's ok.

A few more suggestions would be welcome.
