### 2024-03-07 #6449 @danielskatz
👋 @mikemahoney218, @gonsie, @Chilipp - would one or two of you be willing to review this JOSS submission?
