### 2024-03-07 dd5cb48
Slightly improve inverted permutation as we know we are using integer indices and a unique set of indices. Move import to top of file.
