### 2024-03-07 #6449 @Chilipp
I must admit that I have not yet heard about Dolphin, but I do have some experience in parallelization. If you think that is enough for being a reviewer @jorgensd, I am happy to serve as a reviewer here!
