### 2024-03-07 #6449 @jorgensd
> I must admit that I have not yet heard about Dolphin, but I do have some experience in parallelization. If you think that is enough for being a reviewer @jorgensd, I am happy to serve as a reviewer here!

The software is a MPI based framework for storing checkpoints (function or mesh data) from simulations from FEniCS, that is a software for solving PDEs with finite elements, and making it compatible across runs with different number of processes.

Most of the actual source code is pure numpy (vectorized) and using MPI neighborhood communicators to transfer these arrays to and from processes, along with using ADIOS2.

I have no issue with you being a reviewer.
