### 2024-03-07 #6449 @danielskatz
FYI, the author doesn't actually choose or accept reviewers; this is the role of the editor

But discussion about reviewer suitability is great to have - thanks!
