### 2024-03-09 9a677fe
Remove expensive python loop for dof transformations by using new python binding in DOLFINx (#92)
* Remove expensive python loop for dof transformations by moving it to DOLFINx C++ interface, see: https://github.com/FEniCS/dolfinx/pull/3090

* Ruff formatting

* Slightly improve inverted permutation as we know we are using integer indices and a unique set of indices.
Move import to top of file.

* Apply suggestions from code review
