+++ .github/workflows/test_package_openmpi.yml
@@ -52,6 +52,7 @@ jobs:
         with:
           adios2: ${{ matrix.adios2 }}
           petsc_arch: ${{ env.PETSC_ARCH }}
+          dolfinx:  main
 
       - name: Download legacy data
         uses: actions/download-artifact@v4
