### 2024-03-09 101d865
Remove flake8 and add more info to readme
