### 2024-03-09 45203a4
Update readme and remove flake8 (#94)
* Remove expensive python loop for dof transformations by moving it to DOLFINx C++ interface, see: https://github.com/FEniCS/dolfinx/pull/3090

* Ruff formatting

* Slightly improve inverted permutation as we know we are using integer indices and a unique set of indices.
Move import to top of file.

* Remove flake8 and add more info to readme

* Move contributing guidelines

* ADd joss in review badge
