### 2024-03-25 jorgensd/adios4dolfinx#95 @Chilipp
ipyparallel missing in installation instructions
hey @jorgensd! I am just reviewing your software as part of https://github.com/openjournals/joss-reviews/issues/6451 and I am trying to install it. The conda installation works well, but in your documentation you heavily make use of the [ipyparallel](https://ipyparallel.readthedocs.io/en/latest/) library, that is not mentioned in your installation instructions. this needs to be added
