### 2024-03-25 jorgensd/adios4dolfinx#95 @jorgensd
IPython parallel is listed under optional dependencies:
https://github.com/jorgensd/adios4dolfinx/blob/main/pyproject.toml#L13-L17
I've made a clarification in the installation instructions: https://github.com/jorgensd/adios4dolfinx/pull/96
Let me know if that is a sufficient clarification.

Please note that it is v0.8.0 of this software that is under review.
Unfortunately we do not have a conda release compatible with 0.8.0 yet (as DOLFINx hasn't had a 0.8.0 release, and the current version of the code relies on some interface changes in DOLFINx that I introduced).

Therefore, I would advice you to use `docker` for the time being, for instance with:
` docker run -ti -v $(pwd):/root/shared -w /root/shared --rm ghcr.io/fenics/dolfinx/dolfinx:nightly`
