### 2024-03-25 0371865
Update README.md conda installation instructions (#96)
