### 2024-04-11 a8808ab
refactor mesh creation from reading data
