### 2024-04-11 5be0891
simplify static typing
