### 2024-04-14 jorgensd/adios4dolfinx#95 @Chilipp
@jorgensd unfortunately your fix in #96 did not solve the problem, as you did install `v0.7.3` in the instructions. this version however does not list `ipyparallel` in the [optional dependencies](https://github.com/jorgensd/adios4dolfinx/blob/v0.7.3/pyproject.toml) and as far as I understood, it's not the version we are supposed to review :wink: so please reopen this issue
