### 2024-04-14 jorgensd/adios4dolfinx#95 @jorgensd
I’m a bit uncertain as to what you would like me to specify?
The version under review is 0.8 (which would be equivalent to the main branch, as i will make a new release (0.8.1) once the review is finalized).

Could you elaborate on what you think the current issue (with the main branch) of the software is?
