### 2024-04-14 jorgensd/adios4dolfinx#95 @Chilipp
ah, my apologies @jorgensd, I copied the installation instruction for the test setup from the conda section, and not from the docker section. In docker, you indeed referenced the `main` branch which is ok (see #97)
