### 2024-04-14 5546ced
Refactor mesh creation from reading data (#99)
* refactor mesh creation from reading data

---------

Co-authored-by: nate-sime <>
