### 2024-04-14 #6451 @Chilipp
@danielskatz @jorgensd I am done with the review. Thanks a lot for your great and transparent work on this! Overall I could follow all steps outlined in the documentation (using the docker setup, `conda` was not possible, see https://github.com/jorgensd/adios4dolfinx/issues/95) and it was sufficiently explained. I only have some minor comments that I'd like to share with you in the list below. If you like @jorgensd, I can open independent issues for each of the items if that makes it easier for you to address them. I also mentioned some items, where I was not 100% sure, but checked them off nevertheless. This is just for transparency reasons.

### Problematic checklist items

> - [ ] Contribution and authorship: Has the submitting author (@jorgensd) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

@jorgensd there is another contributor in the [list of contributors on GitHub](https://github.com/jorgensd/adios4dolfinx/graphs/contributors), `@finsberg`. It's hard to judge as an external, whether he should be offered co-authorship here, but as he contributed about 10% of the commits to the repo, please state why he is not listed as a co-author in this paper (only in the Acknowledgements)

> - [ ]  A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?

This is currently missing in the README and documentation.

> - [x] Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

The authors include several guides in the docs with example code to demonstrate the functionality. It's not real-world analysis, but that's because the software is rather generic, so that's ok.

> - [ ] Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?

The authors provide a test-suite that is implemented via CI :+1: Unfortunately, it is missing in the docs, how to run the test suite locally.

> - [ ] Summary: Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?

No, not really. It is not clear to a non-specialist from the summary, what the authors understand under a *checkpointing framework*, what *DOLFINx* is all about, even what *MPI* is should be explained. The summary is using too many technical terms in the summary, that are not obvious for non-specialists.

> - [ ] State of the field: Do the authors describe how this software compares to other commonly-used packages?

As I am not in the field, it's hard to judge this information. The authors however provide some review on how this has been implemented in older versions.

> > Functionality for N-to-M checkpointing was implemented for the old version of DOLFIN by27
(Habera et al., 2018). However, this functionality is not present in the newest version of28
the FEniCS Project (Baratta et al., 2023). The storage principles in the ADIOS4DOLFINx29
are based on the ideas present in this implementation. However, the implementation for30
non-Lagrangian finite element spaces vastly differs, due to the usage of dof-permutations31
(Scroggs et al., 2022). Additionally, all global MPI-calls in the old implementation have been32
reimplemented with scalable MPI-communication using the MPI-3 Neighborhood Collectives33
(MPI-Forum, 2012).

To me, this sounds good enough, but I'd like to leave it to you @danielskatz to checkoff this item, as I am not sure whether I am qualified to evaluate this information.
