### 2024-04-14 1c9d069
Add test instructions
