+++ joss-paper/paper.md
@@ -20,9 +20,12 @@ bibliography: paper.bib
 # Summary
 
 We introduce a checkpointing framework for the latest version of the FEniCS project, known as DOLFINx.
-The framework leverages the data-centric approach of DOLFINx along with a state of the art adaptable Input/Output system called ADIOS2.
-Several variations of checkpointing are supported, including *N-to-M* checkpointing of function data, storage of mesh partitioning information for *N-to-N* checkpointing and snapshot checkpointing for RAM reduction during simulation.
-All MPI operations are using MPI-3 Neighborhood collectives.
+DOLFINx is a general framework for solving partial differential equations using the finite element method.
+The input to simulations using the finite element method is the computational domain (mesh), mesh markers, initial conditions and boundary conditions.
+To be able to restart a simulation at any point, one has to have the capability to read in all of the aforementioned variables.
+The adios4dolfinx package implements all of these operations, using the Message Passing Interface (MPI) for communication across multiple processes and ADIOS2 to for writing/reading data to/from file.
+In particular, the functionality of adios4dolfinx includes "N-to-M"-checkpointing, which means that one can store a result of a simulation that was generated with N number of processes,
+and read it into a program running on M processes.
 
 # Statement of need
 
