### 2024-04-14 #6451 @jorgensd
@Chilipp I've opened various pull requests trying to address most of your comments above:

> > * [ ]   A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?
>
> This is currently missing in the README and documentation.
>

- https://github.com/jorgensd/adios4dolfinx/pull/101

> > * [x]  Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).
>
> The authors include several guides in the docs with example code to demonstrate the functionality. It's not real-world analysis, but that's because the software is rather generic, so that's ok.

- https://github.com/jorgensd/adios4dolfinx/pull/100

> > * [ ]  Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?
>
> The authors provide a test-suite that is implemented via CI 👍 Unfortunately, it is missing in the docs, how to run the test suite locally.

- https://github.com/jorgensd/adios4dolfinx/pull/102

> > * [ ]  Summary: Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?
>
> No, not really. It is not clear to a non-specialist from the summary, what the authors understand under a _checkpointing framework_, what _DOLFINx_ is all about, even what _MPI_ is should be explained. The summary is using too many technical terms in the summary, that are not obvious for non-specialists.
>

- https://github.com/jorgensd/adios4dolfinx/pull/103

> ### Problematic checklist items
>
> > * [ ]  Contribution and authorship: Has the submitting author (@jorgensd) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?
>
> @jorgensd there is another contributor in the [list of contributors on GitHub](https://github.com/jorgensd/adios4dolfinx/graphs/contributors), `@finsberg`. It's hard to judge as an external, whether he should be offered co-authorship here, but as he contributed about 10% of the commits to the repo, please state why he is not listed as a co-author in this paper (only in the Acknowledgements)
>

I'll let @finsberg add his thoughts here. Prior to submission we did discuss if he should be a co-author or not.
