### 2024-04-14 #6451 @jorgensd
> Functionality for N-to-M checkpointing was implemented for the old version of DOLFIN by27
> (Habera et al., 2018). However, this functionality is not present in the newest version of28
> the FEniCS Project (Baratta et al., 2023). The storage principles in the ADIOS4DOLFINx29
> are based on the ideas present in this implementation. However, the implementation for30
> non-Lagrangian finite element spaces vastly differs, due to the usage of dof-permutations31
> (Scroggs et al., 2022). Additionally, all global MPI-calls in the old implementation have been32
> reimplemented with scalable MPI-communication using the MPI-3 Neighborhood Collectives33
> (MPI-Forum, 2012).

I also mention how this has been done in Firedrake and reference their latest preprint: https://arxiv.org/pdf/2401.05868.pdf which wrongly states that legacy DOLFIN and DOLFINx does not support N-to-M checkpointing.
However, they do also state that other popular frameworks such as Moose and Deal ii does not have N-to-M checkpointing implemented.
