### 2024-04-14 71f84b3
Update summary of paper for a non-expert audience (#103)
* Update summary for a non-specialist audience

* Update date on manuscritp
