+++ joss-paper/paper.md
@@ -12,7 +12,7 @@ authors:
 affiliations:
  - name: Simula Research Laboratory
    index: 1
-date: 6 March 2024
+date: 14 April 2024
 bibliography: paper.bib
 
 ---
