### 2024-04-14 9043b82
Add test instructions (#102)
