### 2024-04-14 #6451 @finsberg
> I'll let @finsberg add his thoughts here. Prior to submission we did discuss if he should be a co-author or not.

Hi @Chilipp. @jorgensd and I discussed co-authorship prior to the submission and decided that it was most appropriate to not include me as a co-author. Although I have contributed about 10% of the commits, I don't think these commits have been critical to the software.
