### 2024-04-14 #6451 @danielskatz
> > * [ ]  State of the field: Do the authors describe how this software compares to other commonly-used packages?
>
> As I am not in the field, it's hard to judge this information. The authors however provide some review on how this has been implemented in older versions.
>
> > > Functionality for N-to-M checkpointing was implemented for the old version of DOLFIN by27
> > > (Habera et al., 2018). However, this functionality is not present in the newest version of28
> > > the FEniCS Project (Baratta et al., 2023). The storage principles in the ADIOS4DOLFINx29
> > > are based on the ideas present in this implementation. However, the implementation for30
> > > non-Lagrangian finite element spaces vastly differs, due to the usage of dof-permutations31
> > > (Scroggs et al., 2022). Additionally, all global MPI-calls in the old implementation have been32
> > > reimplemented with scalable MPI-communication using the MPI-3 Neighborhood Collectives33
> > > (MPI-Forum, 2012).
>
> To me, this sounds good enough, but I'd like to leave it to you @danielskatz to checkoff this item, as I am not sure whether I am qualified to evaluate this information.

If it sounds good enough to you, then please check it off.
