### 2024-04-20 fab43a9
Fix bug in reading legacy mesh. Should check that file exist before opening it
