### 2024-04-23 jorgensd/adios4dolfinx#106 @gonsie
problem with installation and running tests
This is related to openjournals/joss-reviews#6451.

I'm trying to install using the docker instructions, found on [this page](http://jsdokken.com/adios4dolfinx/README.html). I've been able to get a dolfinx image running, and install an adios4dolfinx package, but I can't run any successful tests.

I've run:
```
docker run -ti dolfinx/dolfinx:stable
```

Then within the container I have tried various ways to install adios4dolfinx:
```
python3 -m pip install adios4dolfinx[test]
python3 -m pip install adios4dolfinx[test]@git+https://github.com/jorgensd/adios4dolfinx@v0.7.3
```

Both still require me to manually install ipyparallel (which seems like I shouldn't have to from the instructions).

But installing with pip doesn't quite get me to a test environment, so I've separately cloned the repo and tried running the tests:

```
git clone https://github.com/jorgensd/adios4dolfinx.git
cd adios4dolfinx
python3 -m pytest .
```

gives the following errors:
```
python3 -m pytest .
============================================================= test session starts =============================================================
rootdir: /root/adios4dolfinx
configfile: pyproject.toml
plugins: xdist-3.3.1
collected 1145 items / 1 error

=================================================================== ERRORS ====================================================================
_____________________________________________ ERROR collecting tests/test_snapshot_checkpoint.py ______________________________________________
ImportError while importing test module '/root/adios4dolfinx/tests/test_snapshot_checkpoint.py'.
Traceback:
tests/test_snapshot_checkpoint.py:12: in <module>
    from adios4dolfinx.adios2_helpers import resolve_adios_scope
E   ImportError: cannot import name 'resolve_adios_scope' from 'adios4dolfinx.adios2_helpers' (/usr/local/lib/python3.10/dist-packages/adios4dolfinx/adios2_helpers.py)
============================================================== warnings summary ===============================================================
../../usr/local/lib/python3.10/dist-packages/ufl/core/ufl_type.py:56
../../usr/local/lib/python3.10/dist-packages/ufl/core/ufl_type.py:56
../../usr/local/lib/python3.10/dist-packages/ufl/core/ufl_type.py:56
../../usr/local/lib/python3.10/dist-packages/ufl/core/ufl_type.py:56
../../usr/local/lib/python3.10/dist-packages/ufl/core/ufl_type.py:56
../../usr/local/lib/python3.10/dist-packages/ufl/core/ufl_type.py:56
../../usr/local/lib/python3.10/dist-packages/ufl/core/ufl_type.py:56
../../usr/local/lib/python3.10/dist-packages/ufl/core/ufl_type.py:56
  /usr/local/lib/python3.10/dist-packages/ufl/core/ufl_type.py:56: DeprecationWarning: attach_operators_from_hash_data deprecated, please use UFLObject instead.
    warnings.warn("attach_operators_from_hash_data deprecated, please use UFLObject instead.", DeprecationWarning)

-- Docs: https://docs.pytest.org/en/stable/how-to/capture-warnings.html
=========================================================== short test summary info ===========================================================
ERROR tests/test_snapshot_checkpoint.py
!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! Interrupted: 1 error during collection !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
======================================================== 8 warnings, 1 error in 0.15s =========================================================
root@141d6cc482a7:~/adios4dolfinx#
```

I appreciate your assistance.
