### 2024-04-23 #6451 @gonsie
I'm having a lot of trouble with the installation. I'm not sure if it's my (practically nonexistent) docker skills, or poor instructions, but I can't seem to get a full installation.

I've made an issue at jorgensd/adios4dolfinx#106, hopefully with some help I can make progress.
