### 2024-04-23 #6451 @gonsie
One comment about the paper: I've always understood the term "N-to-M checkpointing" to indicate that N processes are writing to M files. With the adios4dolfinx package, are users in control of how many files are used for a checkpoint, or is it always a single shared file?
