+++ src/adios4dolfinx/checkpointing.py
@@ -396,12 +396,8 @@ def read_function(
         dofmap_sorted_by_input = recv_array[unrolled_dofmap_position]
 
         # First invert input data to reference element then transform to current mesh
-        element.pre_apply_transpose_dof_transformation(
-            dofmap_sorted_by_input, input_perms_sorted, bs
-        )
-        element.pre_apply_inverse_transpose_dof_transformation(
-            dofmap_sorted_by_input, inc_perms, bs
-        )
+        element.Tt_apply(dofmap_sorted_by_input, input_perms_sorted, bs)
+        element.Tt_inv_apply(dofmap_sorted_by_input, inc_perms, bs)
         # Compute invert permutation
         inverted_perm = np.empty_like(unrolled_dofmap_position)
         inverted_perm[unrolled_dofmap_position] = np.arange(
