### 2024-04-23 #6451 @jorgensd
> One comment about the paper: I've always understood the term "N-to-M checkpointing" to indicate that N processes are writing to M files. With the adios4dolfinx package, are users in control of how many files are used for a checkpoint, or is it always a single shared file?

I am using the N-to-M checkpointing definition from  https://arxiv.org/pdf/2401.05868.pdf.
The `bp` format is a format made by the developers of ADIOS2, and uses a folder to encapsulate a single binary file, plus some metadata files (https://adios2.readthedocs.io/en/latest/engines/engines.html#bp5).
If you choose to use the backend engine `HDF5` you get a single `.h5` file.
