### 2024-04-23 6347e43
API updates (#107)
* API updates relating to:
https://github.com/FEniCS/dolfinx/pull/3149

* Updates due to: https://github.com/FEniCS/dolfinx/pull/3119
