### 2024-04-23 jorgensd/adios4dolfinx#106 @jorgensd
As stated in https://github.com/jorgensd/adios4dolfinx/issues/95#issuecomment-2053950519 the version under review is 0.8.0 of the software, compatible with dolfinx nightly (as the 0.8.0 is not released, although it is a few days away). Once it is released one should be able to use the new `dolfinx/dolfinx:stable` image, or conda once binaries land there.

There were some API changes over the last few days in the main branch, which I have addressed in: https://github.com/jorgensd/adios4dolfinx/pull/107. These have just been merged into the main branch.

Here is the way I tested this locally:
```bash
docker pull  ghcr.io/fenics/dolfinx/dolfinx:nightly
docker run -ti -v $(pwd):/root/shared -w /root/shared --rm ghcr.io/fenics/dolfinx/dolfinx:nightly
```
Then either install adios4dolfinx with:
```bash
python3 -m pip install adios4dolfinx[test]@git+https://github.com/jorgensd/adios4dolfinx@main
```
or
```bash
python3 -m pip install -e .[test]
```
if you have the repository locally (remember to pull the main branch for the latest changes added 10 minutes ago).

Then run tests with
```bash
python3 -m pytest -xvs .
```
or
```bash
mpirun -n 2 python3 -m pytest -xvs .
```
