### 2024-04-23 jorgensd/adios4dolfinx#106 @jorgensd
@gonsie Let me know if this clarifies and resolves your issue:)
