+++ .github/workflows/test_package.yml
@@ -17,7 +17,6 @@ on:
     - cron: "0 8 * * *"
 
 jobs:
-
   create-datasets:
     uses: ./.github/workflows/create_legacy_data.yml
     with:
@@ -27,7 +26,7 @@ jobs:
     uses: ./.github/workflows/create_legacy_checkpoint.yml
     with:
       artifact_name: "legacy_checkpoint_mpich"
-  
+
   check-formatting:
     uses: ./.github/workflows/check_formatting.yml
 
@@ -49,13 +48,13 @@ jobs:
         with:
           name: legacy_mpich
           path: ./legacy
-    
+
       - name: Download legacy data
         uses: actions/download-artifact@v4
         with:
           name: legacy_checkpoint_mpich
           path: ./legacy_checkpoint
-    
+
       - name: Install package
         run: python3 -m pip install .[test]
 
@@ -65,7 +64,7 @@ jobs:
 
       - name: Run tests in parallel
         run: |
-          mpirun -n 2 coverage run --rcfile=.coveragerc -m mpi4py -m pytest -xvs ./tests/
+          mpirun -n 4 coverage run --rcfile=.coveragerc -m mpi4py -m pytest -xvs ./tests/
 
       - name: Combine coverage reports
         run: |
