### 2024-04-24 73b5012
Try bumping num procs for openmpi as well
