+++ .github/workflows/test_package_openmpi.yml
@@ -74,4 +74,4 @@ jobs:
 
       - name: Run tests in parallel
         run: |
-          mpirun -n 4 coverage run --rcfile=.coveragerc -m mpi4py -m pytest -xvs ./tests/
+          mpirun -n 4 --map-by :OVERSUBSCRIBE coverage run --rcfile=.coveragerc -m mpi4py -m pytest -xvs ./tests/
