### 2024-04-24 b7eb4a4
Increase MPICH runners to 4 processes (#108)
* Increase MPICH runners to 4 processes

* Try bumping num procs for openmpi as well

* Try oversubscribing

* Use environment variable
