### 2024-04-25 jorgensd/adios4dolfinx#106 @gonsie
Thanks, with the nightly build I have been able to run the parallel and serial versions of the tests.

Not sure if you're expecting all tests to pass... I got these results:
- Serial: 158 failed, 1010 passed, 5 skipped in 39.41s
- Parallel: 133 failed, 855 passed, 185 skipped in 31.99s

Thanks.
