+++ .github/workflows/check_formatting.yml
@@ -9,7 +9,7 @@ jobs:
       DEB_PYTHON_INSTALL_LAYOUT: deb_system
 
     runs-on: ubuntu-22.04
-    container: ghcr.io/fenics/dolfinx/dolfinx:nightly
+    container: ghcr.io/fenics/dolfinx/dolfinx:v0.8.0
     steps:
       # This action sets the current path to the root of your github repo
       - uses: actions/checkout@v4
