### 2024-04-25 jorgensd/adios4dolfinx#106 @jorgensd
> Thanks, with the nightly build I have been able to run the parallel and serial versions of the tests.
>
> Not sure if you're expecting all tests to pass... I got these results:
>
>     * Serial: 158 failed, 1010 passed, 5 skipped in 39.41s
>
>     * Parallel: 133 failed, 855 passed, 185 skipped in 31.99s
>
>
> Thanks.

I am expecting all tests to pass. You can see that they are executed on CI at every run:
https://github.com/jorgensd/adios4dolfinx/actions/runs/8829098883/job/24239460728]

I just re-ran the commands I posted in the previous post, and they all pass for me.
Could you post what commands you executed, and what the traceback of one of the failing tests are?
For instance by running `python3 -m pytest -xvs .`
