### 2024-04-25 jorgensd/adios4dolfinx#106 @gonsie
I ran the following:

```
docker run -ti dolfinx/dolfinx:nightly
python3 -m pip install adios4dolfinx[test]@git+https://github.com/jorgensd/adios4dolfinx@main
git clone https://github.com/jorgensd/adios4dolfinx.git
python3 -m pytest -vs .
mpirun -n 2 python3 -m pytest -vs .
```

The first failed serial run was:
```
__________________________________________ test_read_write_P_3D[mesh_3D0-read_comm0-4-Lagrange-True] __________________________________________

read_comm = <mpi4py.MPI.Intracomm object at 0x1555543b4e30>, family = 'Lagrange', degree = 4, is_complex = True
mesh_3D = <dolfinx.mesh.Mesh object at 0x15553301f580>, get_dtype = <function get_dtype.<locals>._get_dtype at 0x155533107eb0>
write_function = <function write_function.<locals>._write_function at 0x15553bc5cca0>
read_function = <function read_function.<locals>._read_function at 0x155533309cf0>

    @pytest.mark.parametrize("is_complex", [True, False])
    @pytest.mark.parametrize("family", ["Lagrange", "DG"])
    @pytest.mark.parametrize("degree", [1, 4])
    @pytest.mark.parametrize("read_comm", [MPI.COMM_SELF, MPI.COMM_WORLD])
    def test_read_write_P_3D(
        read_comm, family, degree, is_complex, mesh_3D, get_dtype, write_function, read_function
    ):
        mesh = mesh_3D
        f_dtype = get_dtype(mesh.geometry.x.dtype, is_complex)
        el = basix.ufl.element(
            family,
            mesh.ufl_cell().cellname(),
            degree,
            basix.LagrangeVariant.gll_warped,
            shape=(mesh.geometry.dim,),
        )

        def f(x):
            values = np.empty((3, x.shape[1]), dtype=f_dtype)
            values[0] = np.pi + x[0]
            values[1] = x[1] + 2 * x[0]
            values[2] = np.cos(x[2])
            if is_complex:
                values[0] -= 2j * x[2]
                values[2] += 1j * x[1]
            return values

        hash = write_function(mesh, el, f, f_dtype)

        MPI.COMM_WORLD.Barrier()
>       read_function(read_comm, el, f, hash, f_dtype)

tests/test_checkpointing.py:105:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _
tests/conftest.py:68: in _read_function
    np.testing.assert_allclose(v.x.array, v_ex.x.array, atol=10 * res, rtol=10 * res)
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

args = (<function assert_allclose.<locals>.compare at 0x1555334c9120>, array([3.94159265+2.77800107e-17j, 1.6       +0.000000...    -1.05453137e-17j, ..., 3.19159265-1.90000000e+00j,
       1.        +0.00000000e+00j, 0.58168309+9.00000000e-01j]))
kwds = {'equal_nan': True, 'err_msg': '', 'header': 'Not equal to tolerance rtol=1e-14, atol=1e-14', 'verbose': True}

    @wraps(func)
    def inner(*args, **kwds):
        with self._recreate_cm():
>           return func(*args, **kwds)
E           AssertionError:
E           Not equal to tolerance rtol=1e-14, atol=1e-14
E
E           Mismatched elements: 16830 / 27783 (60.6%)
E           Max absolute difference: 8016162.74067225
E           Max relative difference: 800859.99683645
E            x: array([3.941593+2.778001e-17j, 1.6     +0.000000e+00j,
E                  1.      -1.054531e-17j, ..., 3.191593-1.900000e+00j,
E                  1.      +0.000000e+00j, 0.581683+9.000000e-01j])
E            y: array([3.941593+2.778001e-17j, 1.6     +0.000000e+00j,
E                  1.      -1.054531e-17j, ..., 3.191593-1.900000e+00j,
E                  1.      +0.000000e+00j, 0.581683+9.000000e-01j])

/usr/lib/python3.10/contextlib.py:79: AssertionError
=========================================================== short test summary info ===========================================================
FAILED tests/test_checkpointing.py::test_read_write_P_3D[mesh_3D0-read_comm0-4-Lagrange-True] - AssertionError:
!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! stopping after 1 failures !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
======================================================== 1 failed, 262 passed in 3.24s ========================================================
```

first failed parallel test:
```
__________ test_read_write_P_2D[mesh_2D3-read_comm0-4-Lagrange-True] ___________

read_comm = <mpi4py.MPI.Intracomm object at 0x155554414490>, family = 'Lagrange'
degree = 4, is_complex = True
mesh_2D = <dolfinx.mesh.Mesh object at 0x155532ed47c0>
get_dtype = <function get_dtype.<locals>._get_dtype at 0x1555450ed5a0>
write_function = <function write_function.<locals>._write_function at 0x155532e2f760>
read_function = <function read_function.<locals>._read_function at 0x155532e2e050>

    @pytest.mark.parametrize("is_complex", [True, False])
    @pytest.mark.parametrize("family", ["Lagrange", "DG"])
    @pytest.mark.parametrize("degree", [1, 4])
    @pytest.mark.parametrize("read_comm", [MPI.COMM_SELF, MPI.COMM_WORLD])
    def test_read_write_P_2D(
        read_comm, family, degree, is_complex, mesh_2D, get_dtype, write_function, read_function
    ):
        mesh = mesh_2D
        f_dtype = get_dtype(mesh.geometry.x.dtype, is_complex)

        el = basix.ufl.element(
            family,
            mesh.ufl_cell().cellname(),
            degree,
            basix.LagrangeVariant.gll_warped,
            shape=(mesh.geometry.dim,),
            dtype=mesh.geometry.x.dtype,
        )

        def f(x):
            values = np.empty((2, x.shape[1]), dtype=f_dtype)
            values[0] = np.full(x.shape[1], np.pi) + x[0]
            values[1] = x[0]
            if is_complex:
                values[0] += 1j * x[1]
                values[1] -= 3j * x[1]
            return values

        hash = write_function(mesh, el, f, f_dtype)
        MPI.COMM_WORLD.Barrier()
>       read_function(read_comm, el, f, hash, f_dtype)

tests/test_checkpointing.py:72:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _
tests/conftest.py:68: in _read_function
    np.testing.assert_allclose(v.x.array, v_ex.x.array, atol=10 * res, rtol=10 * res)
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

args = (<function assert_allclose.<locals>.compare at 0x155532e2f370>, array([3.14159265e+00+0.9j       , 2.77555756e-18-2.7j...5e+00+1.j        , ...,  9.82732684e-01-0.15j      ,
        4.12432534e+00+0.08273268j,  9.82732684e-01-0.24819805j]))
kwds = {'equal_nan': True, 'err_msg': '', 'header': 'Not equal to tolerance rtol=1e-14, atol=1e-14', 'verbose': True}

    @wraps(func)
    def inner(*args, **kwds):
        with self._recreate_cm():
>           return func(*args, **kwds)
E           AssertionError:
E           Not equal to tolerance rtol=1e-14, atol=1e-14
E
E           Mismatched elements: 78 / 3362 (2.32%)
E           Max absolute difference: 10755.05379197
E           Max relative difference: 599.98195738
E            x: array([3.141593e+00+0.9j     , 2.775558e-18-2.7j     ,
E                  3.141593e+00+1.j      , ..., 9.827327e-01-0.15j    ,
E                  4.124325e+00+0.082733j, 9.827327e-01-0.248198j])
E            y: array([ 3.141593e+00+0.9j     , -2.775558e-18-2.7j     ,
E                   3.141593e+00+1.j      , ...,  9.827327e-01-0.15j    ,
E                   4.124325e+00+0.082733j,  9.827327e-01-0.248198j])

/usr/lib/python3.10/contextlib.py:79: AssertionError
=========================== short test summary info ============================
FAILED tests/test_checkpointing.py::test_read_write_P_2D[mesh_2D3-read_comm0-4-Lagrange-True]
!!!!!!!!!!!!!!!!!!!!!!!!!! stopping after 1 failures !!!!!!!!!!!!!!!!!!!!!!!!!!!
======================== 1 failed, 102 passed in 9.61s =========================
```
