### 2024-04-25 jorgensd/adios4dolfinx#106 @jorgensd
I cannot reproduce those error messages running the exact commands you sketched above:
```bash
root@6efd03467dae:~/adios4dolfinx# history
    1  python3 -m pip install adios4dolfinx[test]@git+https://github.com/jorgensd/adios4dolfinx@main
    2  git clone https://github.com/jorgensd/adios4dolfinx.git
    3  cd adios4dolfinx
    4  python3 -m pytest -vs .
    5  mpirun -n 2 python3 -m pytest -vs .
    6  history
```
Could you print the output of:
```bash
 python3 -c "import dolfinx; print(dolfinx.git_commit_hash)"
```
Also what kind of system are you running on?
