### 2024-04-25 jorgensd/adios4dolfinx#106 @nate-sime
I also cannot reproduce those error messages using the commands:

```commandline
docker run -ti dolfinx/dolfinx:nightly
python3 -m pip install adios4dolfinx[test]@git+https://github.com/jorgensd/adios4dolfinx@main
git clone https://github.com/jorgensd/adios4dolfinx.git
cd adios4dolfinx
python3 -m pytest -vs .
mpirun -n 2 python3 -m pytest -vs .
```

In the serial case I get
```
1168 passed, 5 skipped in 50.99s
```
and the parallel case
```
988 passed, 185 skipped in 24.49s
```

For completeness I also see:

```commandline
python3 -c "import dolfinx; print(dolfinx.git_commit_hash)"
d5fb3db2a8544a3e93a3f3aa52d201846b2b9278
```
