### 2024-04-25 jorgensd/adios4dolfinx#106 @gonsie
I see:
```
python3 -c "import dolfinx; print(dolfinx.git_commit_hash)"
d5fb3db2a8544a3e93a3f3aa52d201846b2b9278
```

I'm running the container on a Red Hat Enterprise Linux v8.9 HPC system.
