### 2024-04-25 jorgensd/adios4dolfinx#106 @jorgensd
Very strange... It would be interesting to see if it can be reproduced with conda on your system once:
https://github.com/conda-forge/fenics-dolfinx-feedstock/pull/71
has been merged.

I do unfortunately not have a Red Hat system to try to reproduce this on.
