### 2024-04-25 bc312f5
Remove default build from matrix as redhat image does not have adios2 installed
