+++ .github/workflows/test_redhat.yml
@@ -37,7 +37,7 @@ jobs:
 
     strategy:
       matrix:
-        adios2: ["default", "v2.10.0"]
+        adios2: ["v2.10.0"]
     steps:
       - uses: actions/checkout@v4
 
