### 2024-04-25 20f4286
Re-introduce type hints from typing
