+++ src/adios4dolfinx/legacy_readers.py
@@ -7,6 +7,7 @@
 from __future__ import annotations
 
 import pathlib
+import typing
 
 from mpi4py import MPI
 
@@ -329,7 +330,7 @@ def read_function_from_legacy_h5(
     comm: MPI.Intracomm,
     u: dolfinx.fem.Function,
     group: str = "mesh",
-    step: int | None = None,
+    step: typing.Optional[int] = None,
 ):
     """
     Read function from a `h5`-file generated by legacy DOLFIN `HDF5File.write`
