### 2024-04-25 #6451 @jorgensd
DOLFINx just released a 0.8.0 release (yesterday), and it is making its way onto conda (https://github.com/conda-forge/fenics-dolfinx-feedstock/pull/71).
Hopefully I can get out a 0.8.1 release of the software that aligns with 0.8.0 of DOLFINx (as it would be easier for a user to read installation notes for the software then).
