### 2024-04-25 da2d31c
minor changes in paper.md
