+++ joss-paper/paper.md
@@ -19,38 +19,38 @@ bibliography: paper.bib
 
 # Summary
 
-We introduce a checkpointing framework for the latest version of the FEniCS project, known as DOLFINx.
+We introduce ADIOS4DOLFINx, a checkpointing framework for the latest version of the FEniCS project, known as DOLFINx.
 DOLFINx is a general framework for solving partial differential equations using the finite element method.
-The input to simulations using the finite element method is the computational domain (mesh), mesh markers, initial conditions and boundary conditions.
+The input to simulations using the finite element method is the computational domain (mesh), mesh markers, initial conditions, and boundary conditions.
 To be able to restart a simulation at any point, one has to have the capability to read in all of the aforementioned variables.
-The adios4dolfinx package implements all of these operations, using the Message Passing Interface (MPI) for communication across multiple processes and ADIOS2 to for writing/reading data to/from file.
+The adios4dolfinx package implements all of these operations, using the Message Passing Interface (MPI) for communication across multiple processes and ADIOS2 for writing/reading data to/from file.
 In particular, the functionality of adios4dolfinx includes *N-to-M*-checkpointing, which means that one can store a result of a simulation that was generated with N number of processes,
 and read it into a program running on M processes.
 
 # Statement of need
 
-The ability to start, stop and resume simulations is becoming increasingly important with the growing use of supercomputers for solving scientific and engineering problems.
-A rising number of large scale problems are deployed on high performance, memory distributed computing systems and users tend to run more demanding simulations.
+The ability to start, stop, and resume simulations is becoming increasingly important with the growing use of supercomputers for solving scientific and engineering problems.
+A rising number of large-scale problems are deployed on high-performance, distributed-memory computing systems and users tend to run more demanding simulations.
 These are often non-linear and time-dependent, which typically amounts to thousands of CPU hours.
-As it might uncover bugs and unphysical solutions, the ability to run parts of the simulation, inspect the result and then resume simulation becomes a key factor to enable efficient development.
-If this is discovered early on, the simulation can be terminated saving the developer time, money and energy-usage.
+As it might uncover bugs and unphysical solutions, the ability to run parts of the simulation, inspect the result, and then resume simulation becomes a key factor to enable efficient development.
+If this is discovered early on, the simulation can be terminated, saving the developer time, money and energy usage.
 
-The proposed framework enables users of the FEniCS project [@Baratta:2023] to store solutions during simulation, and read them in at their convenience to resume simulations at a later stage.
+ADIOS4DOLFINx enables users of the FEniCS project [@Baratta:2023] to store solutions during simulation, and read them in at their convenience to resume simulations at a later stage.
 Several checkpointing methods are implemented, including *N-to-M* checkpointing, which means saving data from a program executed with N processes, and loading it back in on M processes.
 
-Functionality for *N-to-M* checkpointing was implemented for the old version of DOLFIN by [@Habera:2018].
+Functionality for *N-to-M* checkpointing was implemented for the old version of DOLFIN by[@Habera:2018.
 However, this functionality is not present in the newest version of the FEniCS Project [@Baratta:2023].
 The storage principles in the ADIOS4DOLFINx are based on the ideas present in this implementation.
 However, the implementation for non-Lagrangian finite element spaces vastly differs, due to the usage of dof-permutations [@Scroggs:2022].
-Additionally, all global MPI-calls in the old implementation have been reimplemented with scalable MPI-communication using the MPI-3 Neighborhood Collectives [@MPI-Forum:2012].
[…]
