### 2024-04-25 0e87d64
Update joss-paper/paper.md
Co-authored-by: Jørgen Schartum Dokken <dokken92@gmail.com>
