### 2024-04-25 #6451 @danielskatz
> > My suggested changes to the paper are in [jorgensd/adios4dolfinx#112](https://github.com/jorgensd/adios4dolfinx/pull/112). Please merge this, or let me know what you disagree with.
>
> All looks good except one citation where i had some suggestions. Thanks for the speedy corrections.

I fixed that one now - thanks for catching it
