### 2024-04-25 51fe02f
Try setting working directory in test
