+++ .github/workflows/test_redhat.yml
@@ -58,21 +58,23 @@ jobs:
         uses: actions/download-artifact@v4
         with:
           name: legacy_ompi
-          path: ./legacy
+          path: ./tests/legacy
 
       - name: Download legacy data
         uses: actions/download-artifact@v4
         with:
           name: legacy_checkpoint_ompi
-          path: ./legacy_checkpoint
+          path: ./tests/legacy_checkpoint
 
       - name: Install package
         run: python3 -m pip install .[test]
 
       - name: Run tests
+        working-directory: ./tests
         run: |
-          coverage run --rcfile=.coveragerc -m mpi4py -m pytest -xvs ./tests/
+          coverage run --rcfile=.coveragerc -m mpi4py -m pytest -xvs .
 
       - name: Run tests in parallel
+        working-directory: ./tests
         run: |
-          mpirun -n 4 coverage run --rcfile=.coveragerc -m mpi4py -m pytest -xvs ./tests/
+          mpirun -n 4 coverage run --rcfile=.coveragerc -m mpi4py -m pytest -xvs .
