### 2024-04-25 2f71f4e
Remove petsc arch
