### 2024-04-25 4dd3b55
try fixing petsc dir again
