### 2024-04-26 305721a
Add notes about adios2 installation
