+++ README.md
@@ -3,7 +3,6 @@
 ![MIT](https://img.shields.io/github/license/jorgensd/adios4dolfinx)
 [![status](https://joss.theoj.org/papers/7866cb142db8a803e32d79a109573d25/status.svg)](https://joss.theoj.org/papers/7866cb142db8a803e32d79a109573d25)
 
-
 ADIOS4DOLFINx is an extension for [DOLFINx](https://github.com/FEniCS/dolfinx/) to checkpoint meshes, meshtags and functions using [ADIOS 2](https://adios2.readthedocs.io/en/latest/).
 
 The code uses the ADIOS2 Python-wrappers to write DOLFINx objects to file, supporting N-to-M (_recoverable_) and N-to-N (_snapshot_) checkpointing.
@@ -12,40 +11,53 @@ See: [Checkpointing in DOLFINx - FEniCS 23](https://jsdokken.com/checkpointing-p
 For scalability, the code uses [MPI Neighbourhood collectives](https://www.mpi-forum.org/docs/mpi-3.1/mpi31-report/node200.htm) for communication across processes.
 
 ## Statement of Need
+
 As the usage of high performance computing clusters increases, more and more large-scale, long-running simulations are deployed.
 The need for storing intermediate solutions from such simulations are crucial, as the HPC system might crash, or the simulation might crash or exceed the alloted computational budget.
 Having a checkpoint of related variables, such as the solutions to partial differential equations (PDEs) is therefore essential.
 The `adios4dolfinx` library extends the [DOLFINx](https://github.com/FEniCS/dolfinx/) computational framework for solving PDEs with checkpointing functionality, such that immediate solutions and mesh information can be stored and re-used in another simulation.
 
-
 ## Installation
+
 Compatibility with DOLFINx:
+
 - ADIOS4DOLFINx v0.7.3 is compatible with DOLFINx v0.7.x
 - ADIOS4DOLFINx v0.8.1 is compatible with DOLFINx v0.8.x
 
+### Dependencies
+
+The library depends on the Python-interface of [DOLFINx](https://github.com/) and an MPI-build of [ADIOS2](https://adios2.readthedocs.io/en/latest/setting_up/setting_up.html#as-package) and can therefore not be installed through PyPi, but has to be installed through Conda, Spack or from source.
+
 ### Docker
 
-ADIOS2 is installed in the official DOLFINx containers, and thus there are no additional dependencies required to install `adios4dolfinx`
+An MPI build of ADIOS2 is installed in the official DOLFINx containers, and thus there are no additional dependencies required to install `adios4dolfinx`
 on top of DOLFINx in these images.
 
 Create a Docker container, named for instance `dolfinx-checkpoint`.
-Use the `nightly` tag to get the main branch of DOLFINx, or `stable` to get the lastest stable release
+Use the `nightly` tag to get the main branch of DOLFINx, or `stable` to get the latest stable release
+
 ```bash
 docker run -ti -v $(pwd):/root/shared -w /root/shared --name=dolfinx-checkpoint ghcr.io/fenics/dolfinx/dolfinx:nightly
 ```
+
 For the latest version compatible with nightly (with the ability to run the test suite), use
+
 ```bash
 python3 -m pip install adios4dolfinx[test]@git+https://github.com/jorgensd/adios4dolfinx@main
 ```
+
 If you are using the `stable` image, you can install `adios4dolfinx` from [PYPI](https://pypi.org/project/adios4dolfinx/) with
+
 ```bash
 python3 -m pip install adios4dolfinx[test]
 ```
 
 This docker container can be opened with
+
 ```bash
[…]
