+++ .github/workflows/test_redhat.yml
@@ -36,6 +36,7 @@ jobs:
       PETSC_ARCH: ""
       PETSC_DIR: "/usr/local/"
       PYTHONPATH: "/usr/local/lib/:${PYTHONPATH}"
+      working-directory: ./src
 
     strategy:
       matrix:
@@ -47,7 +48,7 @@ jobs:
         run: python3 -m pip install --upgrade pip setuptools
 
       - name: Install DOLFINx
-        uses: jorgensd/actions/install-dolfinx@v0.2.0
+        uses: jorgensd/actions/install-dolfinx@v0.3.0
         with:
           adios2: ${{ matrix.adios2 }}
           petsc_arch: ${{ env.PETSC_ARCH }}
@@ -56,28 +57,27 @@ jobs:
           basix: v0.8.0
           ufl: 2024.1.0.post0
           ffcx: v0.8.0
+          working-directory: ${{ env.working-directory}}
 
       - name: Download legacy data
         uses: actions/download-artifact@v4
         with:
           name: legacy_ompi
-          path: ./tests/legacy
+          path: ./legacy
 
       - name: Download legacy data
         uses: actions/download-artifact@v4
         with:
           name: legacy_checkpoint_ompi
-          path: ./tests/legacy_checkpoint
+          path: ./legacy_checkpoint
 
       - name: Install package
         run: python3 -m pip install .[test]
 
       - name: Run tests
-        working-directory: ./tests
         run: |
-          coverage run --rcfile=.coveragerc -m mpi4py -m pytest -xvs .
+          coverage run --rcfile=.coveragerc -m mpi4py -m pytest -xvs ./tests
 
       - name: Run tests in parallel
-        working-directory: ./tests
         run: |
-          mpirun -n 4 coverage run --rcfile=.coveragerc -m mpi4py -m pytest -xvs .
+          mpirun -n 4 coverage run --rcfile=.coveragerc -m mpi4py -m pytest -xvs ./tests
