+++ pyproject.toml
@@ -8,7 +8,7 @@ description = "Checkpointing functionality for DOLFINx meshes/functions with ADI
 authors = [{ name = "Jørgen S. Dokken", email = "dokken@simula.no" }]
 license = { file = "LICENSE" }
 readme = "README.md"
-dependencies = ["fenics-dolfinx>=0.8.0", "adios2"]
+dependencies = ["fenics-dolfinx>=0.8.0"]
 
 [project.optional-dependencies]
 test = ["pytest", "coverage", "ipyparallel"]
