### 2024-04-26 7bf967c
Test redhat installation (#111)
* Test package on redhat

* Remove default build from matrix as redhat image does not have adios2 installed

* Re-introduce type hints from typing

* Add back adios2 to the Python dependencies

* Try setting working directory in test

* Remove petsc arch

* try fixing petsc dir again

* Install adios2 from source as it requires MPI build. Set pythonpath in parent workflow

* Add notes about adios2 installation

* Use newest install-dolfinx action

* Remove adios2 from deps as the pypi version is incompatible with package (no MPI support)
