### 2024-04-26 jorgensd/adios4dolfinx#106 @jorgensd
What MPI is installed on your HPC system?
