### 2024-04-26 jorgensd/adios4dolfinx#106 @gonsie
I'm using the [poodle system](https://hpc.llnl.gov/hardware/compute-platforms/poodle) at LLNL, intel CPUs. Of note, I'm also using podman rather than docker directly, so there may be some issues there. That or an MPI issue would be the first places I would look. However, I'm not really interested in debugging this further. I'm happy to share the list of tests that are failing, if that would be helpful.
