### 2024-04-26 jorgensd/adios4dolfinx#106 @jorgensd
It would be interesting to find out what is wrong, but it is unlikely that i will have a similar system to replicate that setup on. Therefore I’ll close this issue.
