### 2024-04-27 20895da
v0.8.1 release (#110)
* Revert 0.8 API changes for release

* Test against 0.7.3 image

* Update citation file and pyproject.toml

* Update workflows

* Correct version in readme

* Test redhat installation (#111)

* Test package on redhat

* Remove default build from matrix as redhat image does not have adios2 installed

* Re-introduce type hints from typing

* Add back adios2 to the Python dependencies

* Try setting working directory in test

* Remove petsc arch

* try fixing petsc dir again

* Install adios2 from source as it requires MPI build. Set pythonpath in parent workflow

* Add not
[…]
