+++ pyproject.toml
@@ -3,12 +3,12 @@ requires = ["setuptools>=61.0.0", "wheel"]
 
 [project]
 name = "adios4dolfinx"
-version = "0.8.0.dev0"
+version = "0.8.1"
 description = "Checkpointing functionality for DOLFINx meshes/functions with ADIOS2"
 authors = [{ name = "Jørgen S. Dokken", email = "dokken@simula.no" }]
 license = { file = "LICENSE" }
 readme = "README.md"
-dependencies = ["fenics-dolfinx>=0.8.0.dev0"]
+dependencies = ["fenics-dolfinx>=0.8.0"]
 
 [project.optional-dependencies]
 test = ["pytest", "coverage", "ipyparallel"]
