### 2024-04-27 jorgensd/adios4dolfinx#113 @jorgensd
Most tests pass locally.
Currently ipyparallel tests fail with:
```bash
___________ test_read_write_3D_vector_non_simplex[4-NCF-False-False] ___________

write_mesh = False, family = 'NCF', degree = 4, is_complex = False
create_non_simplex_mesh_3D = PosixPath('/tmp/pytest-of-conda/pytest-0/output11/original_mesh_3D_non_simplex.xdmf')
cluster = <ipyparallel.client.client.Client object at 0x7fc86e847ca0>
get_dtype = <function get_dtype.<locals>._get_dtype at 0x7fc86e545700>
tmp_path = PosixPath('/tmp/pytest-of-conda/pytest-0/test_read_write_3D_vector_non_7')

    @pytest.mark.skipif(MPI.COMM_WORLD.size > 1, reason="Test uses ipythonparallel for MPI")
    @pytest.mark.parametrize("write_mesh", [True, False])
    @pytest.mark.parametrize("is_complex", [True, False])
    @pytest.mark.parametrize("family", ["NCF"])
    @pytest.mark.parametrize("degree", [1, 4])
    def test_read_write_3D_vector_non_simplex(
        write_mesh, family, degree, is_complex, create_non_simplex_mesh_3D, cluster, get_dtype, tmp_path
    ):
        fname = create_non_simplex_mesh_3D

        f_dtype = get_dtype(np.float64, is_complex)

        def f(x):
            values = np.empty((3, x.shape[1]), dtype=f_dtype)
            values[0] = np.full(x.shape[1], np.pi) + x[0]
            values[1] = np.cos(x[2])
            values[2] = x[0]
            if is_complex:
                values[2] += x[0] * x[1] * 3j
            return values

        query = cluster[:].apply_async(
            write_function_vector, write_mesh, fname, family, degree, f, f_dtype, "u_original", tmp_path
        )
        query.wait()
>       assert query.successful(), query.error
E       AssertionError: [<RemoteError[ ]:ModuleNotFoundError(No module named 'test_original_checkpoint')>, <RemoteError[ ]:ModuleNotFoundError(No module named 'test_original_checkpoint')>]
E       assert False
E        +  where False = <bound method AsyncResult.successful of <AsyncResult(write_function_vector): failed>>()
E        +    where <bound method AsyncResult.successful of <AsyncResult(write_function_vector): failed>> = <AsyncResult(write_function_vector): failed>.successful

tests/test_original_checkpoint.py:500: AssertionError
```
