### 2024-04-29 #6451 @jorgensd
> 👋 @jorgensd - note that I'm waiting for you to do the items in [#6451 (comment)](https://github.com/openjournals/joss-reviews/issues/6451#issuecomment-2078115312) ...

Hi Daniel, I'm currently working on getting a conda build up and running, and would like to do that prior to tagging the repository and making a release. Sorry for the inconvenience. I've made an issue at: https://github.com/jorgensd/adios4dolfinx/issues/113 tracking my progress wrt. this endeavor.
