### 2024-04-29 7e3832c
ADd debug func
