### 2024-04-30 07ba3be
TST: sync sys.path on engines before tests
ensures imports resolve the same for engines

pytest injects test dir to sys.path, while engines might not receive that
