### 2024-04-30 e7c41db
Add Min RK to acknowledgement section of paper
