+++ joss-paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'ADIOS4DOLFINx: A framework for checkpointing in FEniCS'
+title: "ADIOS4DOLFINx: A framework for checkpointing in FEniCS"
 tags:
   - Python
   - finite element simulations
@@ -10,11 +10,10 @@ authors:
     corresponding: true
     affiliation: 1
 affiliations:
- - name: Simula Research Laboratory
-   index: 1
-date: 14 April 2024
+  - name: Simula Research Laboratory
+    index: 1
+date: 30 April 2024
 bibliography: paper.bib
-
 ---
 
 # Summary
@@ -24,7 +23,7 @@ DOLFINx is a general framework for solving partial differential equations using 
 The input to simulations using the finite element method is the computational domain (mesh), mesh markers, initial conditions, and boundary conditions.
 To be able to restart a simulation at any point, one has to have the capability to read in all of the aforementioned variables.
 The adios4dolfinx package implements all of these operations, using the Message Passing Interface (MPI) for communication across multiple processes and ADIOS2 for writing/reading data to/from file.
-In particular, the functionality of adios4dolfinx includes *N-to-M*-checkpointing, which means that one can store a result of a simulation that was generated with N number of processes,
+In particular, the functionality of adios4dolfinx includes _N-to-M_-checkpointing, which means that one can store a result of a simulation that was generated with N number of processes,
 and read it into a program running on M processes.
 
 # Statement of need
@@ -84,7 +83,7 @@ A large variety of examples covering all the functions in ADIOS4DOLFINx is avail
 
 # Acknowledgements
 
-We acknowledge the valuable feedback on the documentation and manuscript by Thomas M. Surowiec and Halvor Herlyng. 
+We acknowledge the valuable feedback on the documentation and manuscript by Thomas M. Surowiec and Halvor Herlyng and packaging support by Min Ragan-Kelley.
 Additionally, we acknowledge the scientific discussion regarding feature development and code contributions by Francesco Ballarin, Henrik N. Finsberg, and Nathan Sime.
 
 # References
