### 2024-04-30 59500c9
Add Min RK to acknowledgement section of paper (#115)
* Add Min RK to acknowledgement section of paper

* Revert autoformatting
