### 2024-04-30 1c07ab0
Add installation notes for conda. Not tested yet (#116)
