### 2024-04-30 d719425
Expand adios2 build matrix on openmpi to ensure we are testing 2.9.2 and 2.10.0 (#117)
* Expand adios2 build matrix on openmpi to ensure we are testing 2.9.2 and 2.10.0

* Bump version number
