+++ .github/workflows/test_package_openmpi.yml
@@ -41,7 +41,7 @@ jobs:
 
     strategy:
       matrix:
-        adios2: ["default", "v2.10.0"]
+        adios2: ["default", "v2.9.2", "v2.10.0"]
     steps:
       - uses: actions/checkout@v4
 
